requirements.txt
torch==1.13.0
torchvision==0.14.0
grpcio-tools
grpcio
fastapi
uvicorn[standard]
grequests
aiohttp
furl
starlette-prometheus
grpcio
grpcio-tools
supervisor